Export the menu and logo to be a separate AppHeader component that can be imported into all the app's screens
