+++ src/containers/intro/IntroSection.js
@@ -4,49 +4,16 @@
 
 import React from 'react';
 import styled from 'styled-components';
-import { Link } from 'react-router-dom';
 
-import logoImg from '../../assets/images/logo-white.png';
-
+import AppHeader from '../../components/headers/AppHeader';
 import BackgroundCarousel from './BackgroundCarousel';
 import PageSection from '../../components/layout/PageSection';
 import StyledSectionSubTitle from '../../components/headers/StyledSectionSubTitle';
 import StyledSectionTitle from '../../components/headers/StyledSectionTitle';
 import { N0, PK0 } from '../../core/style/Colors';
-import { WINDOW_EDGE_PADDING, MEDIA_QUERY_MD, MEDIA_QUERY_LG } from '../../core/style/Sizes';
-import * as Routes from '../../core/router/Routes';
+import { MEDIA_QUERY_MD, MEDIA_QUERY_LG } from '../../core/style/Sizes';
 
-const helpLink = 'https://help.openlattice.com';
 const scheduleACallLink = 'https://calendly.com/openlattice/openlattice-web-request';
-
-const Header = styled.div`
-  display: flex;
-  flex: 1 0 auto;
-  justify-content: space-between;
-  margin-top: ${WINDOW_EDGE_PADDING}px;
-`;
-
-const Menu = styled.div`
-  align-items: center;
-  display: flex;
-  font-size: 14px;
-  justify-content: space-between;
-`;
-
-const MenuInternalLink = styled(Link)`
-  color: ${N0};
-  margin-left: 30px;
-  text-decoration: none;
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-const MenuExternalLink = styled.a`
-  color: ${N0};
-  margin-left: 30px;
-  text-decoration: none;
-`;
 
 const Content = styled.div`
   color: ${N0};
@@ -114,13 +81,7 @@ const BackgroundComponent = (
 
 const IntroSection = () => (
   <PageSection bgComponent={BackgroundComponent}>
-    <Header>
-      <img src={logoImg} alt="OpenLattice Logo" height={50} />
-      <Menu>
-        <MenuInternalLink to={Routes.PRODUCTS}>Products</MenuInternalLink>
-        <MenuExternalLink href={helpLink}>Help</MenuExternalLink>
-      </Menu>
-    </Header>
+    <AppHeader />
     <Content>
       <Title>
         Modern data infrastructure for core societal institutions
